Started implementing conditional rendering based on neighborhood member status
