+++ apps/frontend/src/pages/individual_neighborhood/components/DescriptionBox/DescriptionBox.module.css
