+++ apps/frontend/src/pages/individual_neighborhood/components/MemberBox/MemberBox.module.css
