+++ apps/frontend/src/pages/individual_neighborhood/components/RequestBox/RequestBox.module.css
