+++ apps/frontend/src/pages/individual_neighborhood/neighborhood.module.css
