+++ apps/frontend/src/pages/login/login.module.css
@@ -1,0 +1,8 @@
+.loginContainer {
+  height: 100%;
+  display: flex; 
+  align-items: center;
+  position:absolute;
+  top: 0; bottom: 0; left: 0; right: 0;
+  margin: auto;
+}
