+++ apps/frontend/src/pages/private_routes/privateRoutes.tsx
@@ -1,9 +1,10 @@
-import { Navigate, Outlet, useOutletContext } from 'react-router-dom';
+import { Navigate, Outlet } from 'react-router-dom';
 import React from 'react';
 
 const PrivateRoutes = () => {  
   //@ts-ignore
   let auth = JSON.parse(window.localStorage.getItem('user'));
+  console.log(auth);
   const [userContext, setUserContext] = React.useState(auth);
   if (auth) {
       if (auth.token) {
@@ -12,6 +13,7 @@ const PrivateRoutes = () => {
         return (<Navigate to='/login'/>)
       }
   }
+
   return <Navigate to='/login' />
 }
