--- apps/frontend/src/routes/login/login.module.css
@@ -1,5 +1,0 @@
-.loginContainer {
-  height: 100%;
-  display: flex; 
-  align-items: center;
-}
